chore: extract remoteHandler
