+++ .changeset/wise-trains-live.md
@@ -1,0 +1,5 @@
+---
+'@module-federation/runtime': patch
+---
+
+chore: extract remoteHandler
