+++ packages/runtime/src/core.ts
@@ -1,8 +1,6 @@
-import type { ModuleInfo, GlobalModuleInfo } from '@module-federation/sdk';
+import type { ModuleInfo } from '@module-federation/sdk';
 import {
   Options,
-  PreloadAssets,
-  PreloadOptions,
   PreloadRemoteArgs,
   RemoteEntryExports,
   Remote,
@@ -16,9 +14,7 @@ import {
 } from './type';
 import {
   assert,
-  warn,
   getBuilderId,
-  addUniqueItem,
   matchRemoteWithNameAndExpose,
   registerPlugins,
 } from './utils';
@@ -30,16 +26,14 @@ import {
   SyncHook,
   SyncWaterfallHook,
 } from './utils/hooks';
-import { getGlobalShareScope } from './utils/share';
-import { formatPreloadArgs, preloadAssets } from './utils/preload';
 import { generatePreloadAssetsPlugin } from './plugins/generate-preload-assets';
 import { snapshotPlugin } from './plugins/snapshot';
 import { isBrowserEnv } from './utils/env';
-import { getRemoteEntryUniqueKey, getRemoteInfo } from './utils/load';
-import { globalLoading } from './global';
-import { DEFAULT_REMOTE_TYPE, DEFAULT_SCOPE } from './constant';
+import { getRemoteInfo } from './utils/load';
+import { DEFAULT_SCOPE } from './constant';
 import { SnapshotHandler } from './plugins/snapshot/SnapshotHandler';
 import { SharedHandler } from './shared';
+import { RemoteHandler } from './remote';
 
 export interface LoadRemoteMatch {
   id: string;
@@ -64,24 +58,26 @@ export async function getRemoteModuleAndOptions(options: {
   let loadRemoteArgs;
 
   try {
-    loadRemoteArgs = await origin.hooks.lifecycle.beforeRequest.emit({
-      id,
-      options: origin.options,
-      origin: origin,
-    });
+    loadRemoteArgs =
+      await origin.remoteHandler.hooks.lifecycle.beforeRequest.emit({
+        id,
+        options: origin.options,
+        origin: origin,
+      });
   } catch (error) {
-    loadRemoteArgs = (await origin.hooks.lifecycle.errorLoadRemote.emit({
-      id,
-      options: origin.options,
-      origin: origin,
-      from: 'runtime',
-      error,
-      lifecycle: 'beforeRequest',
-    })) as {
-      id: string;
-      options: Options;
-      origin: FederationHost;
-    };
+    loadRemoteArgs =
+      (await origin.remoteHandler.hooks.lifecycle.errorLoadRemote.emit({
+        id,
+        options: origin.options,
+        origin: origin,
+        from: 'runtime',
+        error,
+        lifecycle: 'beforeRequest',
+      })) as {
+        id: string;
+        options: Options;
+        origin: FederationHost;
+      };
 
     if (!loadRemoteArgs) {
       throw error;
@@ -165,11 +161,6 @@ export class FederationHost {
       ],
       void
     >(),
-    beforeRequest: new AsyncWaterfallHook<{
-      id: string;
-      options: Options;
-      origin: FederationHost;
-    }>('beforeRequest'),
     // maybe will change, temporarily for internal use only
     beforeInitContainer: new AsyncWaterfallHook<{
       shareScope: ShareScopeMap[string];
@@ -187,76 +178,13 @@ export class FederationHost {
       remoteEntryExports: RemoteEntryExports;
       origin: FederationHost;
     }>('initContainer'),
-    onLoad: new AsyncHook<
-      [
-        {
-          id: string;
-          expose: string;
-          pkgNameOrAlias: string;
-          remote: Remote;
-          options: ModuleOptions;
-          origin: FederationHost;
-          exposeModule: any;
-          exposeModuleFactory: any;
-          moduleInstance: Module;
-        },
-      ],
-      void
-    >('onLoad'),
-    handlePreloadModule: new SyncHook<
-      {
-        id: string;
-        name: string;
-        remote: Remote;
-        remoteSnapshot: ModuleInfo;
-        preloadConfig: PreloadRemoteArgs;
-        origin: FederationHost;
-      },
-      void
-    >('handlePreloadModule'),
-    errorLoadRemote: new AsyncHook<
-      [
-        {
-          id: string;
-          error: unknown;
-          options?: any;
-          from: 'build' | 'runtime';
-          lifecycle: 'onLoad' | 'beforeRequest';
-          origin: FederationHost;
-        },
-      ],
-      void | unknown
-    >('errorLoadRemote'),
-    beforePreloadRemote: new AsyncHook<{
-      preloadOps: Array<PreloadRemoteArgs>;
-      options: Options;
-      origin: FederationHost;
-    }>(),
-    generatePreloadAssets: new AsyncHook<
-      [
-        {
-          origin: FederationHost;
-          preloadOptions: PreloadOptions[number];
-          remote: Remote;
-          remoteInfo: RemoteInfo;
-          remoteSnapshot: ModuleInfo;
-          globalSnapshot: GlobalModuleInfo;
-        },
-      ],
-      Promise<PreloadAssets>
-    >('generatePreloadAssets'),
-    // not used yet
-    afterPreloadRemote: new AsyncHook<{
-      preloadOps: Array<PreloadRemoteArgs>;
-      options: Options;
-      origin: FederationHost;
-    }>(),
   });
   version: string = __VERSION__;
   name: string;
   moduleCache: Map<string, Module> = new Map();
   snapshotHandler: SnapshotHandler;
   sharedHandler: SharedHandler;
+  remoteHandler: RemoteHandler;
   shareScopeMap: ShareScopeMap;
   loaderHook = new PluginSystem({
     // FIXME: may not be suitable , not open to the public yet
@@ -308,20 +236,13 @@ export class FederationHost {
     this.options = defaultOptions;
     this.snapshotHandler = new SnapshotHandler(this);
     this.sharedHandler = new SharedHandler(this.options);
+    this.remoteHandler = new RemoteHandler();
     this.shareScopeMap = this.sharedHandler.shareScopeMap;
     this.registerPlugins([
       ...defaultOptions.plugins,
       ...(userOptions.plugins || []),
     ]);
     this.options = this.formatOptions(defaultOptions, userOptions);
-  }
-
-  private _setGlobalShareScopeMap(): void {
-    const globalShareScopeMap = getGlobalShareScope();
-    const identifier = this.options.id || this.options.name;
-    if (identifier && !globalShareScopeMap[identifier]) {
-      globalShareScopeMap[identifier] = this.shareScopeMap;
-    }
   }
 
   initOptions(userOptions: UserOptions): Options {
@@ -384,93 +305,12 @@ export class FederationHost {
     id: string,
     options?: { loadFactory?: boolean; from: 'build' | 'runtime' },
   ): Promise<T | null> {
-    try {
-      const { loadFactory = true } = options || { loadFactory: true };
-      // 1. Validate the parameters of the retrieved module. There are two module request methods: pkgName + expose and alias + expose.
-      // 2. Request the snapshot information of the current host and globally store the obtained snapshot information. The retrieved module information is partially offline and partially online. The online module information will retrieve the modules used online.
-      // 3. Retrieve the detailed information of the current module from global (remoteEntry address, expose resource address)
-      // 4. After retrieving remoteEntry, call the init of the module, and then retrieve the exported content of the module through get
-      // id: pkgName(@federation/app1) + expose(button) = @federation/app1/button
-      // id: alias(app1) + expose(button) = app1/button
-      // id: alias(app1/utils) + expose(loadash/sort) = app1/utils/loadash/sort
-      const { module, moduleOptions, remoteMatchInfo } =
-        await getRemoteModuleAndOptions({
-          id,
-          origin: this,
-        });
-      const { pkgNameOrAlias, remote, expose, id: idRes } = remoteMatchInfo;
-      const moduleOrFactory = (await module.get(expose, options)) as T;
-
-      const moduleWrapper = await this.hooks.lifecycle.onLoad.emit({
-        id: idRes,
-        pkgNameOrAlias,
-        expose,
-        exposeModule: loadFactory ? moduleOrFactory : undefined,
-        exposeModuleFactory: loadFactory ? undefined : moduleOrFactory,
-        remote,
-        options: moduleOptions,
-        moduleInstance: module,
-        origin: this,
-      });
-
-      if (typeof moduleWrapper === 'function') {
-        return moduleWrapper as T;
-      }
-
-      return moduleOrFactory;
-    } catch (error) {
-      const { from = 'runtime' } = options || { from: 'runtime' };
-
-      const failOver = await this.hooks.lifecycle.errorLoadRemote.emit({
-        id,
-        error,
-        from,
-        lifecycle: 'onLoad',
-        origin: this,
-      });
-
-      if (!failOver) {
-        throw error;
-      }
-
-      return failOver as T;
-    }
+    return this.remoteHandler.loadRemote(this, id, options);
   }
 
   // eslint-disable-next-line @typescript-eslint/member-ordering
   async preloadRemote(preloadOptions: Array<PreloadRemoteArgs>): Promise<void> {
-    await this.hooks.lifecycle.beforePreloadRemote.emit({
-      preloadOptions,
-      options: this.options,
-      origin: this,
-    });
-
-    const preloadOps: PreloadOptions = formatPreloadArgs(
-      this.options.remotes,
-      preloadOptions,
-    );
-
-    await Promise.all(
-      preloadOps.map(async (ops) => {
-        const { remote } = ops;
-        const remoteInfo = getRemoteInfo(remote);
-        const { globalSnapshot, remoteSnapshot } =
-          await this.snapshotHandler.loadRemoteSnapshotInfo(remote);
-
-        const assets = await this.hooks.lifecycle.generatePreloadAssets.emit({
-          origin: this,
-          preloadOptions: ops,
-          remote,
-          remoteInfo,
-          globalSnapshot,
-          remoteSnapshot,
-        });
-        if (!assets) {
-          return;
-        }
-        preloadAssets(remoteInfo, this, assets);
-      }),
-    );
+    return this.remoteHandler.preloadRemote(this, preloadOptions);
   }
 
   initShareScopeMap(
@@ -496,12 +336,11 @@ export class FederationHost {
         shareInfo: shared,
       });
 
-    const userRemotes = userOptionsRes.remotes || [];
-
-    const remotes = userRemotes.reduce((res, remote) => {
-      this.registerRemote(remote, res, { force: false });
-      return res;
-    }, globalOptionsRes.remotes);
+    const remotes = this.remoteHandler.formatRemote(
+      this,
+      globalOptionsRes,
+      userOptionsRes,
+    );
 
     this.sharedHandler.registerShared(shareInfos, userOptions);
 
@@ -533,6 +372,7 @@ export class FederationHost {
   registerPlugins(plugins: UserOptions['plugins']) {
     const pluginRes = registerPlugins(plugins, [
       this.hooks,
+      this.remoteHandler.hooks,
       this.sharedHandler.hooks,
       this.snapshotHandler.hooks,
       this.loaderHook,
@@ -547,96 +387,7 @@ export class FederationHost {
     }, pluginRes || []);
   }
 
-  private removeRemote(remote: Remote): void {
-    const { name } = remote;
-    const remoteIndex = this.options.remotes.findIndex(
-      (item) => item.name === name,
-    );
-    if (remoteIndex !== -1) {
-      this.options.remotes.splice(remoteIndex, 1);
-    }
-    const loadedModule = this.moduleCache.get(remote.name);
-    if (loadedModule) {
-      const key = loadedModule.remoteInfo
-        .entryGlobalName as keyof typeof globalThis;
-      if (globalThis[key]) {
-        delete globalThis[key];
-      }
-      const remoteEntryUniqueKey = getRemoteEntryUniqueKey(
-        loadedModule.remoteInfo,
-      );
-      if (globalLoading[remoteEntryUniqueKey]) {
-        delete globalLoading[remoteEntryUniqueKey];
-      }
-      this.moduleCache.delete(remote.name);
-    }
-  }
-
-  private registerRemote(
-    remote: Remote,
-    targetRemotes: Remote[],
-    options?: { force?: boolean },
-  ): void {
-    const normalizeRemote = () => {
-      if (remote.alias) {
-        // Validate if alias equals the prefix of remote.name and remote.alias, if so, throw an error
-        // As multi-level path references cannot guarantee unique names, alias being a prefix of remote.name is not supported
-        const findEqual = targetRemotes.find(
-          (item) =>
-            remote.alias &&
-            (item.name.startsWith(remote.alias) ||
-              item.alias?.startsWith(remote.alias)),
-        );
-        assert(
-          !findEqual,
-          `The alias ${remote.alias} of remote ${
-            remote.name
-          } is not allowed to be the prefix of ${
-            findEqual && findEqual.name
-          } name or alias`,
-        );
-      }
-      // Set the remote entry to a complete path
-      if ('entry' in remote) {
-        if (isBrowserEnv() && !remote.entry.startsWith('http')) {
-          remote.entry = new URL(remote.entry, window.location.origin).href;
-        }
-      }
-      if (!remote.shareScope) {
-        remote.shareScope = DEFAULT_SCOPE;
-      }
-      if (!remote.type) {
-        remote.type = DEFAULT_REMOTE_TYPE;
-      }
-    };
-    const registeredRemote = targetRemotes.find(
-      (item) => item.name === remote.name,
-    );
-    if (!registeredRemote) {
-      normalizeRemote();
-      targetRemotes.push(remote);
-    } else {
-      const messages = [
-        `The remote "${remote.name}" is already registered.`,
-        options?.force
-          ? 'Hope you have known that OVERRIDE it may have some unexpected errors'
-          : 'If you want to merge the remote, you can set "force: true".',
-      ];
-      if (options?.force) {
-        // remove registered remote
-        this.removeRemote(registeredRemote);
-        normalizeRemote();
-        targetRemotes.push(remote);
-      }
-      warn(messages.join(' '));
-    }
-  }
-
   registerRemotes(remotes: Remote[], options?: { force?: boolean }): void {
-    remotes.forEach((remote) => {
-      this.registerRemote(remote, this.options.remotes, {
-        force: options?.force,
-      });
-    });
+    return this.remoteHandler.registerRemotes(this, remotes, options);
   }
 }
