+++ packages/runtime/src/plugins/generate-preload-assets.ts
@@ -213,7 +213,7 @@ export function generatePreloadAssets(
         for (let index = 0; index < assetsLength; index++) {
           const assetsInfo = moduleAssetsInfo[index];
           const exposeFullPath = `${remoteInfo.name}/${assetsInfo.moduleName}`;
-          origin.hooks.lifecycle.handlePreloadModule.emit({
+          origin.remoteHandler.hooks.lifecycle.handlePreloadModule.emit({
             id:
               assetsInfo.moduleName === '.' ? remoteInfo.name : exposeFullPath,
             name: remoteInfo.name,
